+++ src/main/java/com/example/dao/PetRepository.java
@@ -7,4 +7,5 @@ import com.example.domain.PetVO;
 @Repository
 public interface PetRepository extends CrudRepository<PetVO, Integer>{
 
+
 }
