+++ src/main/java/com/example/service/loginService/LoginService.java
@@ -1,7 +1,5 @@
 package com.example.service.loginService;
 
-import java.util.List;
-
 import com.example.domain.ImgVO;
 import com.example.domain.UserVO;
 
@@ -12,6 +10,7 @@ public interface LoginService {
 	public ImgVO findBypRimgname(String pRimgname);
 
 	public void userImgUpdate(ImgVO ivo);
+
 
 //	/* 회원탈퇴용으로 쓰는 중*/
 //	public void delete(UserVO vo);
